Introduce a common type between FXConnectionSkin (renamed to DefaultFXConnectionSkin) and FXNewConnectionSkin
