+++ VWorkflows-FX/src/main/java/eu/mihosoft/vrl/workflow/fx/FXNewConnectionSkin.java
@@ -33,208 +33,66 @@
  */
 package eu.mihosoft.vrl.workflow.fx;
 
-import eu.mihosoft.vrl.workflow.Connection;
 import eu.mihosoft.vrl.workflow.ConnectionResult;
 import eu.mihosoft.vrl.workflow.Connector;
 import eu.mihosoft.vrl.workflow.VFlow;
 import eu.mihosoft.vrl.workflow.VFlowModel;
 import eu.mihosoft.vrl.workflow.VNode;
-import eu.mihosoft.vrl.workflow.skin.ConnectionSkin;
-
-import javafx.beans.binding.DoubleBinding;
-import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleObjectProperty;
 import javafx.event.EventHandler;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.shape.Circle;
 import javafx.scene.shape.CubicCurveTo;
-import javafx.scene.shape.LineTo;
-import javafx.scene.shape.MoveTo;
-import javafx.scene.shape.Path;
 import jfxtras.labs.util.event.MouseControlUtil;
 
 /**
  *
  * @author Michael Hoffer &lt;info@michaelhoffer.de&gt;
  */
-public class FXNewConnectionSkin implements ConnectionSkin<Connection>, FXSkin<Connection, Path> {
-
-    private ObjectProperty<Connector> senderProperty = new SimpleObjectProperty<>();
-    private ObjectProperty<Connector> receiverProperty = new SimpleObjectProperty<>();
-    private Path connectionPath;
-    private LineTo lineTo;
-    private MoveTo moveTo;
+public class FXNewConnectionSkin extends AbstractFXConnectionSkin {
     private CubicCurveTo curveTo = new CubicCurveTo();
-//    private Shape startConnector;
-    private Circle receiverConnectorUI;
     private ConnectorShape senderConnectorUI;
     private VFlowModel flow;
-    private VFlow flowController;
-    private ObjectProperty<Connection> modelProperty = new SimpleObjectProperty<>();
-    private ObjectProperty<Parent> parentProperty = new SimpleObjectProperty<>();
-    private String type;
     private Node lastNode;
-    private FXSkinFactory skinFactory;
-    private ConnectionListener connectionListener;
 
     public FXNewConnectionSkin(FXSkinFactory skinFactory,
             Parent parent, Connector sender, VFlow controller, String type) {
-        this.skinFactory = skinFactory;
-        setParent(parent);
+        super(skinFactory, parent, controller, type);
         setSender(sender);
-
-        this.flowController = controller;
-        this.flow = controller.getModel();
-        this.type = type;
-
-//        startConnector = new Circle(20);
-        receiverConnectorUI = new Circle(8);
-
-        moveTo = new MoveTo();
-        lineTo = new LineTo();
-        connectionPath = new Path(moveTo, curveTo);
-
-        init();
+        flow = controller.getModel();
     }
 
-    private void init() {
-
+    protected void initStyle() {
         connectionPath.getStyleClass().setAll(
-                "vnode-new-connection",
-                "vnode-new-connection-" + getSender().getType());
+            "vnode-new-connection",
+            "vnode-new-connection-" + getSender().getType());
         receiverConnectorUI.getStyleClass().setAll(
-                "vnode-new-connection-receiver",
-                "vnode-new-connection-receiver-" + getSender().getType());
+            "vnode-new-connection-receiver",
+            "vnode-new-connection-receiver-" + getSender().getType());
 
         receiverConnectorUI.setStrokeWidth(3);
-//    
+        //
+    }
+
+    protected void initSenderAndReceiver() {
+        receiverConnectorUI = new Circle(15);
         final VNode sender = getSender().getNode();
         final FXFlowNodeSkin senderSkin = (FXFlowNodeSkin) getController().
                 getNodeSkinLookup().getById(skinFactory, sender.getId());
         
-        final ConnectorShape sendershape = senderSkin.getConnectorShape(getSender());
-        final Node senderNode = sendershape.getNode();
+        senderShape = senderSkin.getConnectorShape(getSender());
+        final Node senderNode = senderShape.getNode();
 
-        senderConnectorUI = sendershape;
-
-        DoubleBinding startXBinding = new DoubleBinding() {
-            {
-                super.bind(senderNode.layoutXProperty(),
-                    sendershape.radiusProperty());
-            }
-
-            @Override
-            protected double computeValue() {
-
-                return senderNode.getLayoutX()
-                        + sendershape.getRadius();
-            }
-        };
-
-        DoubleBinding startYBinding = new DoubleBinding() {
-            {
-                super.bind(senderNode.layoutYProperty(),
-                    sendershape.radiusProperty());
-            }
-
-            @Override
-            protected double computeValue() {
-                return senderNode.getLayoutY()
-                        + sendershape.getRadius();
-            }
-        };
-
-
-        DoubleBinding controlX1Binding = new DoubleBinding() {
-            {
-                super.bind(senderNode.boundsInLocalProperty(),
-                        senderNode.layoutXProperty(), 
-                        receiverConnectorUI.layoutXProperty());
-            }
-
-            @Override
-            protected double computeValue() {
-
-                return senderNode.getLayoutX() 
-                        + (receiverConnectorUI.getLayoutX()
-                        - senderNode.getLayoutX()) / 2;
-
-            }
-        };
-
-        DoubleBinding controlY1Binding = new DoubleBinding() {
-            {
-                super.bind(senderNode.boundsInLocalProperty(),
-                        senderNode.layoutYProperty());
-            }
-
-            @Override
-            protected double computeValue() {
-
-                return senderNode.getLayoutY();
-
-            }
-        };
-
-        DoubleBinding controlX2Binding = new DoubleBinding() {
-            {
-                super.bind(senderNode.boundsInLocalProperty(),
-                        senderNode.layoutXProperty(),
-                        receiverConnectorUI.layoutXProperty());
-            }
-
-            @Override
-            protected double computeValue() {
-
-                return receiverConnectorUI.getLayoutX() 
-                        - (receiverConnectorUI.getLayoutX()
-                        - senderNode.getLayoutX()) / 2;
-
-            }
-        };
-
-        DoubleBinding controlY2Binding = new DoubleBinding() {
-            {
-                super.bind(receiverConnectorUI.boundsInLocalProperty(),
-                        receiverConnectorUI.layoutYProperty());
-            }
-
-            @Override
-            protected double computeValue() {
-
-                return receiverConnectorUI.getLayoutY();
-
-            }
-        };
-
-        moveTo.xProperty().bind(startXBinding);
-        moveTo.yProperty().bind(startYBinding);
-
-        curveTo.controlX1Property().bind(controlX1Binding);
-        curveTo.controlY1Property().bind(controlY1Binding);
-        curveTo.controlX2Property().bind(controlX2Binding);
-        curveTo.controlY2Property().bind(controlY2Binding);
-        curveTo.xProperty().bind(
-                receiverConnectorUI.layoutXProperty());
-        curveTo.yProperty().bind(receiverConnectorUI.layoutYProperty());
-
-        makeDraggable();
+        senderConnectorUI = senderShape;
 
         receiverConnectorUI.setLayoutX(senderNode.getLayoutX()
                 +receiverConnectorUI.getRadius());
         receiverConnectorUI.setLayoutY(senderNode.getLayoutY()
                 +receiverConnectorUI.getRadius());
-
-        connectionListener
-                = new DefaultConnectionListener(
-                        skinFactory, flowController, receiverConnectorUI);
-
     }
 
-    private void makeDraggable() {
-
+    protected void makeDraggable() {
         connectionPath.toFront();
         receiverConnectorUI.toFront();
 
@@ -371,16 +229,6 @@ public class FXNewConnectionSkin implements ConnectionSkin<Connection>, FXSkin<C
             
             remove();
         });
-
-    }
-
-    public Node getReceiverConnector() {
-        return receiverConnectorUI;
-    }
-
-    @Override
-    public Connector getSender() {
-        return senderProperty.get();
     }
 
     @Override
@@ -394,98 +242,8 @@ public class FXNewConnectionSkin implements ConnectionSkin<Connection>, FXSkin<C
     }
 
     @Override
-    public ObjectProperty<Connector> senderProperty() {
-        return senderProperty;
-    }
-
-    @Override
-    public Connector getReceiver() {
-        return receiverProperty.get();
-    }
-
-    @Override
-    public void setReceiver(Connector n) {
-        receiverProperty.set(n);
-    }
-
-    @Override
-    public ObjectProperty<Connector> receiverProperty() {
-        return receiverProperty;
-    }
-
-    @Override
-    public Path getNode() {
-        return connectionPath;
-    }
-
-    @Override
-    public Parent getContentNode() {
-        return getParent();
-    }
-
-    @Override
-    public void setModel(Connection model) {
-        modelProperty.set(model);
-    }
-
-    @Override
-    public Connection getModel() {
-        return modelProperty.get();
-    }
-
-    @Override
-    public ObjectProperty<Connection> modelProperty() {
-        return modelProperty;
-    }
-
-    final void setParent(Parent parent) {
-        parentProperty.set(parent);
-    }
-
-    Parent getParent() {
-        return parentProperty.get();
-    }
-
-    ObjectProperty<Parent> parentProperty() {
-        return parentProperty;
-    }
-
-    @Override
-    public void add() {
-        NodeUtil.addToParent(getParent(), connectionPath);
-        NodeUtil.addToParent(getParent(), receiverConnectorUI);
-
-        receiverConnectorUI.toFront();
-        connectionPath.toFront();
-    }
-
-    @Override
-    public void remove() {
-        NodeUtil.removeFromParent(connectionPath);
-        NodeUtil.removeFromParent(receiverConnectorUI);
-    }
-
-    @Override
-    public VFlow getController() {
-        return flowController;
-    }
-
-    @Override
     public void setController(VFlow flow) {
-        this.flowController = flow;
+        super.setController(flow);
         this.flow = flow.getModel();
-    }
-
-    /**
-     * @return the skinFactory
-     */
-    @Override
-    public FXSkinFactory getSkinFactory() {
-        return skinFactory;
-    }
-
-    @Override
-    public void receiverToFront() {
-        receiverConnectorUI.toFront();
     }
 }
